pipeline/pyproject.toml
[project]
name = "insidertrack-pipeline"
version = "0.1.0"
description = "InsiderTrack data pipeline for CVM insider trading data"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "pdfplumber>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pyright>=1.1.0",
]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"
reportMissingImports = "warning"
